+++ astropy/nddata/tests/test_utils.py
@@ -337,9 +337,10 @@ class TestCutout2D(object):
         self.wcs = wcs
 
     def test_cutout(self):
-        for shape in [(3, 3), (3*u.pixel, 3*u.pix)]:
+        sizes = [3, 3*u.pixel, (3, 3), (3*u.pixel, 3*u.pix), (3., 3*u.pixel)]
+        for size in sizes:
             position = (2.1, 1.9)
-            c = Cutout2D(self.data, position, shape)
+            c = Cutout2D(self.data, position, size)
             assert c.data.shape == (3, 3)
             assert c.data[1, 1] == 10
             assert c.origin_original == (1, 1)
@@ -355,45 +356,43 @@ class TestCutout2D(object):
             assert c.slices_original == (slice(1, 4), slice(1, 4))
             assert c.slices_cutout == (slice(0, 3), slice(0, 3))
 
-    def test_cutout_sidelength(self):
-        for side_length in [3, 3*u.pixel]:
-            position = (2.1, 1.9)
-            c = Cutout2D(self.data, position, side_length=side_length)
-            assert c.data.shape == (3, 3)
-            assert c.data[1, 1] == 10
-            assert c.origin_original == (1, 1)
-            assert c.origin_cutout == (0, 0)
-            assert c.input_position_original == position
-            assert_allclose(c.input_position_cutout, (1.1, 0.9))
-            assert c.position_original == (2., 2.)
-            assert c.position_cutout == (1., 1.)
-            assert c.center_original == (2., 2.)
-            assert c.center_cutout == (1., 1.)
-            assert c.bbox_original == ((1, 3), (1, 3))
-            assert c.bbox_cutout == ((0, 2), (0, 2))
-            assert c.slices_original == (slice(1, 4), slice(1, 4))
-            assert c.slices_cutout == (slice(0, 3), slice(0, 3))
+    def test_size_length(self):
+        with pytest.raises(ValueError):
+            Cutout2D(self.data, (2, 2), (1, 1, 1))
+
+    def test_size_units(self):
+        for size in [3 * u.cm, (3, 3 * u.K)]:
+            with pytest.raises(ValueError):
+                Cutout2D(self.data, (2, 2), size)
+
+    def test_size_angle(self):
+        c = Cutout2D(self.data, (2, 2), (0.1* u.arcsec), wcs=self.wcs)
+        assert c.data.shape == (2, 2)
+        assert c.data[0, 0] == 5
+        assert c.slices_original == (slice(1, 3), slice(1, 3))
+        assert c.slices_cutout == (slice(0, 2), slice(0, 2))
+
+    def test_size_angle_without_wcs(self):
+        with pytest.raises(ValueError):
+            Cutout2D(self.data, (2, 2), (3, 3* u.arcsec))
 
     def test_cutout_trim_overlap(self):
-        shape = (3, 3)
-        c = Cutout2D(self.data, (0, 0), shape, mode='trim')
+        c = Cutout2D(self.data, (0, 0), (3, 3), mode='trim')
         assert c.data.shape == (2, 2)
         assert c.data[0, 0] == 0
         assert c.slices_original == (slice(0, 2), slice(0, 2))
         assert c.slices_cutout == (slice(0, 2), slice(0, 2))
 
     def test_cutout_partial_overlap(self):
-        shape = (3, 3)
-        c = Cutout2D(self.data, (0, 0), shape, mode='partial')
+        c = Cutout2D(self.data, (0, 0), (3, 3), mode='partial')
         assert c.data.shape == (3, 3)
         assert c.data[1, 1] == 0
         assert c.slices_original == (slice(0, 2), slice(0, 2))
         assert c.slices_cutout == (slice(1, 3), slice(1, 3))
 
     def test_cutout_partial_overlap_fill_value(self):
-        shape = (3, 3)
         fill_value = -99
-        c = Cutout2D(self.data, (0, 0), shape, mode='partial',
+        c = Cutout2D(self.data, (0, 0), (3, 3), mode='partial',
                    fill_value=fill_value)
         assert c.data.shape == (3, 3)
         assert c.data[1, 1] == 0
@@ -416,8 +415,7 @@ class TestCutout2D(object):
 
     def test_to_from_large(self):
         position = (2, 2)
-        shape = (3, 3)
-        c = Cutout2D(self.data, position, shape)
+        c = Cutout2D(self.data, position, (3, 3))
         xy = (0, 0)
         result = c.to_cutout_position(c.to_original_position(xy))
         assert_allclose(result, xy)